# Fix type-stability for normalization layers