This PR fixes type-stability for normalization layers.
This also makes ResNet model type-stable (as well as others).

While there is [PR](https://github.com/FluxML/Flux.jl/pull/1509) that reworks normalization layers, it is not clear what is the status of it.
@ToucheSir has also suggested that functional part should be moved into NNlib.jl, so I'm not sure if this PR should be accepted in the first place...
But at least we can use it to look at what improvements type-stability can bring in this case...

~~Also I feel it is not the cleanest solution, as it essentially computes the output type `o::_basetype(typeof(x)){O, N}`.~~

### Benchmark:

```julia
using Test
using Flux
using BenchmarkTools

function main()
    x = rand(Float32, 28, 28, 64, 4)
    bn = BatchNorm(64)
    θ = Flux.params(bn)

    @info "forward"
    @time bn(x)
    @btime $bn($x)

    @info "grad"
    @time gradient(θ) do
        sum(bn(x))
    end
    @btime gradient($θ) do
        sum($bn($x))
    end

    @inferred bn(x) # master fails here
end
```

Before:

```julia
[ Info: forward
  0.254515 seconds (856.55 k allocations: 43.045 MiB, 98.02% compilation time)
  911.059 μs (29 allocations: 785.33 KiB)
[ Info: grad
  4.050493 seconds (12.31 M allocations: 635.845 MiB, 3.65% gc time, 99.81% compilation time)
  2.838 ms (804 allocations: 13.05 MiB)
```

After:

```julia
[ Info: forward
  0.000917 seconds (15 allocations: 1.532 MiB)
  455.609 μs (15 allocations: 1.53 MiB)
[ Info: grad
  2.245501 seconds (7.16 M allocations: 372.776 MiB, 3.75% gc time, 99.17% compilation time)
  2.337 ms (694 allocations: 12.28 MiB)
```

### PR Checklist

- [x] Tests are added
- [ ] Entry in NEWS.md
- [ ] Documentation, if applicable
- [ ] API changes require approval from a committer (different from the author, if applicable)