@pxl-th 2022-02-03
> How do the times look now? More or less the same?

Yes, the times are the same.

```julia
[ Info: forward
  0.000813 seconds (15 allocations: 1.532 MiB)
  461.305 μs (15 allocations: 1.53 MiB)
[ Info: grad
  1.970701 seconds (7.04 M allocations: 366.512 MiB, 4.10% gc time, 99.81% compilation time)
  2.209 ms (634 allocations: 12.28 MiB)
```