@ToucheSir 2022-02-03
Magnificent, thanks!

P.S.
> @ToucheSir has also suggested that functional part should be moved into NNlib.jl, so I'm not sure if this PR should be accepted in the first place...

For anyone reading this comment, here's a "good first issue" ;)